Find start of VDM before checking checksum
check_nmea_checksum was being run on the whole string which may have
other NMEA-like checksums embedded in it. This moves it after the point
where the start of the VDM message is found, and it only runs on the VDM
message part of the string.

This also adds a test string to runme.py and the c test suite.